[Add] Withdraw Page - crypto withdraw amount range

+++ app/src/main/java/org/cxct/sportlottery/network/withdraw/add/WithdrawAddRequest.kt
@@ -1,7 +1,7 @@
 package org.cxct.sportlottery.network.withdraw.add
 
 data class WithdrawAddRequest(
-    val applyMoney: Long,
+    val applyMoney: Double,
     val withdrawPwd: String,
     val id: Long,
 )
